move function from user to cost
