+++ .idea/workspace.xml
@@ -2,10 +2,8 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="67b0171a-bc0d-43f8-a941-f65852c0ce83" name="Changes" comment="">
-      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/public/javascripts/MangeBill.js" beforeDir="false" afterPath="$PROJECT_DIR$/public/javascripts/MangeBill.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/public/javascripts/billdatabase.js" beforeDir="false" afterPath="$PROJECT_DIR$/public/javascripts/billdatabase.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/routes/cost.js" beforeDir="false" afterPath="$PROJECT_DIR$/routes/cost.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/routes/users.js" beforeDir="false" afterPath="$PROJECT_DIR$/routes/users.js" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -63,7 +61,7 @@
       <workItem from="1655968023832" duration="2525000" />
       <workItem from="1656006855773" duration="264000" />
       <workItem from="1656007147009" duration="212000" />
-      <workItem from="1656069312940" duration="1388000" />
+      <workItem from="1656069312940" duration="1772000" />
     </task>
     <servers />
   </component>
